requirements.txt
python-telegram-bot==13.15
urllib3==1.26.15
six==1.16.0
python-dotenv
gspread
oauth2client
requests
Pillow
python-dotenv
